Provide a more useful error when signing up with an invalid name
Resolves #29

I'm editing the app.po file directly, because these two strings seem
locked in Transifex and I can't edit them there.

The proper solution for this would be to change the source text to
include these clarifications and not leave the error message like this.
However, the source of the problem seems to reside in MySociety's
commonlib package and thus the change seems hard to make.

Moreover, some people have less than 3 names legally. Let's hope we
don't get many of them. :)