Outgoing messages should default the address to the info request’s one
The user should be able to overwrite that address for each outgoing
message as mentioned in #15, but if not specified, the address should
default to the one in the info request.

The fixtures, however, can not have a missing address field as they’re
directly mapped to a DB record and loading the fixtures will fail.